requirements.txt
Flask==3.0.3
mysql-connector-python==8.4.0
bcrypt==4.1.3
PyJWT==2.8.0
python-dotenv==1.0.1
